CASP-828 Fix wss for master branch, SNOW-169012 Fix C3861 C2039 'runtime_error' is not a member of 'std'  (#208)

+++ ci/build/wss.sh
@@ -1,22 +1,40 @@
 #!/usr/bin/env bash
 #
-# Run whitesource for NodeJS driver
+# Run whitesource
 #
 set -e
 set -o pipefail
 THIS_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
-
 [[ -z "$WHITESOURCE_API_KEY" ]] && echo "[WARNING] No WHITESOURCE_API_KEY is set. No WhiteSource scan will occur." && exit 0
 
 export PRODUCT_NAME=libsnowflakeclient
-export PROJECT_NAME=libsnowflakeclient
 
-DATE=$(date +'%m-%d-%Y')
+env | grep GITHUB | sort
+
+export PROD_BRANCH=master
+export PROD_GIT_REF=refs/heads/$PROD_BRANCH
+export PROJECT_VERSION=$GITHUB_SHA
+
+if [[ "$GITHUB_EVENT_NAME" == "pull_request" ]]; then
+    echo "[INFO] Pull Request"
+    IFS="/"
+    read -ra GITHUB_REF_ELEMS <<<"$GITHUB_REF"
+    IFS=" "
+    export PROJECT_NAME=PR-${GITHUB_REF_ELEMS[2]}
+elif [[ "$GITHUB_REF" == "$PROD_GIT_REF" ]]; then
+    echo "[INFO] Production branch"
+    export PROJECT_NAME=$PROD_BRANCH
+else
+    echo "[INFO] Non Production branch. Skipping wss..."
+    export PROJECT_NAME=
+fi
 
 SCAN_DIRECTORIES=$(cd $THIS_DIR/../.. && pwd)
 
-rm -f wss-unified-agent.jar 
-curl -LO https://github.com/whitesource/unified-agent-distribution/releases/latest/download/wss-unified-agent.jar
+if [[ -n "$PROJECT_NAME" ]]; then
+    rm -f wss-unified-agent.jar 
+    curl -LO https://github.com/whitesource/unified-agent-distribution/releases/latest/download/wss-unified-agent.jar
+fi
 
 SCAN_CONFIG=wss-libsnowflakeclient-agent.config
 cat > $SCAN_CONFIG <<CONFIG
@@ -51,9 +69,9 @@ wss.url=https://saas.whitesourcesoftware.com/agent
 # Policies #
 ############
 checkPolicies=true
-forceCheckAllDependencies=false
-forceUpdate=false
-forceUpdate.failBuildOnPolicyViolation=false
+forceCheckAllDependencies=true
+forceUpdate=true
+forceUpdate.failBuildOnPolicyViolation=true
 #updateInventory=false
 
 ###########
@@ -98,28 +116,57 @@ case.sensitive.glob=false
 followSymbolicLinks=true
 CONFIG
 
-echo "[INFO] Running wss.sh for ${PROJECT_NAME}-${PRODUCT_NAME} under ${SCAN_DIRECTORIES}"
-java -jar wss-unified-agent.jar -apiKey ${WHITESOURCE_API_KEY} \
-    -c ${SCAN_CONFIG} \
-    -project ${PROJECT_NAME} \
-    -product ${PRODUCT_NAME} \
-    -d ${SCAN_DIRECTORIES} \
-    -wss.url https://saas.whitesourcesoftware.com/agent \
-    -offline true
-
-if java -jar wss-unified-agent.jar -apiKey ${WHITESOURCE_API_KEY} \
-   -c ${SCAN_CONFIG} \
-   -project ${PROJECT_NAME} \
-   -product ${PRODUCT_NAME} \
-   -projectVersion baseline \
-   -requestFiles whitesource/update-request.txt \
-   -wss.url https://saas.whitesourcesoftware.com/agent ; then
-    echo "checkPolicies=false" >> ${SCAN_CONFIG}
+set +e
+if [[ "$PROJECT_NAME" == "$PROD_BRANCH" ]]; then
+    # Prod branch
+    echo "[INFO] Running wss.sh for ${PRODUCT_NAME}-${PROJECT_NAME}-${PROJECT_VERSION} under ${SCAN_DIRECTORIES}"
     java -jar wss-unified-agent.jar -apiKey ${WHITESOURCE_API_KEY} \
         -c ${SCAN_CONFIG} \
-        -project ${PROJECT_NAME} \
+        -d ${SCAN_DIRECTORIES} \
         -product ${PRODUCT_NAME} \
-        -projectVersion ${DATE} \
-        -requestFiles whitesource/update-request.txt \
-        -wss.url https://saas.whitesourcesoftware.com/agent
+        -project ${PROJECT_NAME} \
+        -projectVersion ${PROJECT_VERSION} \
+        -offline true
+    ERR=$?
+    if [[ "$ERR" != "254" && "$ERR" != "0" ]]; then
+        echo "failed to run wss for $PRODUCT_VERSION_${PROJECT_VERSION} in ${PROJECT_VERSION}..."
+        exit 1
+    fi
+    java -jar wss-unified-agent.jar -apiKey ${WHITESOURCE_API_KEY} \
+       -c ${SCAN_CONFIG} \
+       -product ${PRODUCT_NAME} \
+       -project ${PROJECT_NAME} \
+       -projectVersion baseline \
+       -requestFiles whitesource/update-request.txt
+    ERR=$?
+    if [[ "$ERR" != "254" && "$ERR" != "0" ]]; then
+        echo "failed to run wss for $PRODUCT_VERSION_${PROJECT_VERSION} in baseline"
+        exit 1
+    fi
+    java -jar wss-unified-agent.jar -apiKey ${WHITESOURCE_API_KEY} \
+        -c ${SCAN_CONFIG} \
+        -product ${PRODUCT_NAME} \
+        -project ${PROJECT_NAME} \
+        -projectVersion ${PROJECT_VERSION} \
+        -requestFiles whitesource/update-request.txt
+    ERR=$?
+    if [[ "$ERR" != "254" && "$ERR" != "0" ]]; then
+        echo "failed to run wss for $PRODUCT_VERSION_${PROJECT_VERSION} in ${PROJECT_VERSION}"
+        exit 1
+    fi
+elif [[ -n "$PROJECT_NAME" ]]; then
+    # PR
+    echo "[INFO] Running wss.sh for ${PRODUCT_NAME}-${PROJECT_NAME}-${PROJECT_VERSION} under ${SCAN_DIRECTORIES}"
+    java -jar wss-unified-agent.jar -apiKey ${WHITESOURCE_API_KEY} \
+        -c ${SCAN_CONFIG} \
+        -d ${SCAN_DIRECTORIES} \
+        -product ${PRODUCT_NAME} \
+        -project ${PROJECT_NAME} \
+        -projectVersion ${PROJECT_VERSION}
+    ERR=$?
+    if [[ "$ERR" != "254" && "$ERR" != "0" ]]; then
+        echo "failed to run wss for $PRODUCT_VERSION_${PROJECT_VERSION} in ${PROJECT_VERSION}..."
+        exit 1
+    fi
 fi
+set -e
